Fix deletion of users when no mail account is present
Instead of catching a null reference for non-existing mail accounts, it seems
to be easier to only get the mail acccount when we actually want to do
something with it. Since we are only intersted in the change of the display
name, the retreival of the user account has been moved to that if-branch.
This also seems more like a Nextcloud bug, since it fails on errors of hooks.